Add basic DownloadForm component
